Update exercise_rain_threshold.ipynb
Hint and solution were mixtures of this exercise and the previous one where T750 is added as an output variable. Corrected for this, but some additional text is likely needed in the solution where the student can compare the values of a CLM variable (that would be affected by the rain_threshold change) between this run and the run created in the previous exercise. I do not know what variable should be compared.